+++ src/ConnectorRuntimeConfig.ts
@@ -3,6 +3,8 @@ import * as log4js from "log4js";
 import { ConnectorRuntimeModuleConfiguration } from "./ConnectorRuntimeModule";
 
 export interface ConnectorRuntimeConfig extends RuntimeConfig {
+    debug: boolean;
+
     database: {
         connectionString: string;
         dbName: string;