+++ src/ConnectorRuntimeModule.ts
@@ -1,4 +1,6 @@
+import { ILogger } from "@js-soft/logging-abstractions";
 import { ModuleConfiguration, RuntimeModule } from "@nmshd/runtime";
+import { ConnectorMode } from "./ConnectorMode";
 import { ConnectorRuntime } from "./ConnectorRuntime";
 
 export interface ConnectorRuntimeModuleConfiguration extends ModuleConfiguration {
@@ -8,4 +10,8 @@ export interface ConnectorRuntimeModuleConfiguration extends ModuleConfiguration
 export abstract class ConnectorRuntimeModule<TConfig extends ConnectorRuntimeModuleConfiguration = ConnectorRuntimeModuleConfiguration> extends RuntimeModule<
     TConfig,
     ConnectorRuntime
-> {}
+> {
+    public constructor(runtime: ConnectorRuntime, configuration: TConfig, logger: ILogger, protected readonly connectorMode: ConnectorMode) {
+        super(runtime, configuration, logger);
+    }
+}