+++ src/infrastructure/ConnectorInfastructure.ts
@@ -1,4 +1,5 @@
 import { ILogger } from "@js-soft/logging-abstractions";
+import { ConnectorMode } from "../ConnectorMode";
 import { ConnectorRuntime } from "../ConnectorRuntime";
 
 export interface InfrastructureConfiguration {
@@ -6,7 +7,13 @@ export interface InfrastructureConfiguration {
 }
 
 export abstract class ConnectorInfrastructure<TConfig extends InfrastructureConfiguration = InfrastructureConfiguration> {
-    public constructor(protected runtime: ConnectorRuntime, protected configuration: TConfig, protected logger: ILogger, public readonly name: string) {}
+    public constructor(
+        protected runtime: ConnectorRuntime,
+        protected configuration: TConfig,
+        protected logger: ILogger,
+        public readonly name: string,
+        protected readonly connectorMode: ConnectorMode
+    ) {}
 
     public get isEnabled(): boolean {
         return this.configuration.enabled;