+++ src/infrastructure/httpServer/common/Envelope.ts
@@ -1,13 +1,19 @@
-import { HttpError } from "./HttpError";
+import { ConnectorMode } from "../../../ConnectorMode";
+import { HttpError, HttpErrorJSON } from "./HttpError";
 
 export class Envelope {
-    protected constructor(public result?: any, public error?: HttpError) {}
+    protected constructor(public result?: any, public error?: HttpErrorJSON) {}
 
     public static ok(result: any): Envelope {
         return new Envelope(result, undefined);
     }
 
-    public static error(error: HttpError): Envelope {
-        return new Envelope(undefined, error);
+    public static error(error: HttpError, connectorMode: ConnectorMode): Envelope {
+        switch (connectorMode) {
+            case "debug":
+                return new Envelope(undefined, error.toJSON("verbose"));
+            case "production":
+                return new Envelope(undefined, error.toJSON());
+        }
     }
 }